Modified tests a bit. oh god this is so broken
